+++ doc/release-notes.md
@@ -4,4 +4,9 @@ Bitcoin ABC version 0.30.12 is now available from:
 
   <https://download.bitcoinabc.org/0.30.12/>
 
-This is a maintenance release with no user-visible change.
+This release includes the following features and fixes:
+ - The `availability_score` field from the `getavalanchepeerinfo` RPC is
+   deprecated and no longer computed. This field was never used outside of this
+   RPC. To keep the field in the RPC output, add
+   `deprecatedrpc=peer_availability_score` to your node configuration. Note that
+   the value will always be 0.